refactor: Split Transactions.jsx into TransactionRow and TopMost
