+++ src/ducks/transactions/Transactions.jsx
@@ -1,8 +1,6 @@
-import React, { PureComponent } from 'react'
+import React from 'react'
 import ReactDOM from 'react-dom'
-import cx from 'classnames'
 import {
-  find,
   sortBy,
   throttle,
   debounce,
@@ -12,228 +10,24 @@ import {
   groupBy
 } from 'lodash'
 import { format } from 'date-fns'
+import { isIOSApp } from 'cozy-device-helper'
+
 import {
   translate,
   withBreakpoints,
-  Media,
-  Bd,
-  Img,
-  Caption,
-  Text,
   Button
 } from 'cozy-ui/react'
-import { Figure } from 'components/Figure'
 import * as List from 'components/List'
-import { Table, TdSecondary } from 'components/Table'
-import TransactionActions from './TransactionActions'
-import { getLabel } from './helpers'
-import {
-  getAccountLabel,
-  getAccountInstitutionLabel
-} from 'ducks/account/helpers'
-import {
-  getParentCategory,
-  getCategoryName
-} from 'ducks/categories/categoriesMap'
-import { getCategoryId } from 'ducks/categories/helpers'
-import CategoryIcon from 'ducks/categories/CategoryIcon'
-import { withUpdateCategory } from 'ducks/categories'
-import { withDispatch } from 'utils'
+import { Table } from 'components/Table'
+
 import styles from './Transactions.styl'
-import { InfiniteScroll, ScrollRestore } from './scroll'
+import { InfiniteScroll, ScrollRestore, TopMost } from './scroll'
 import TransactionModal from './TransactionModal'
-import { isIOSApp } from 'cozy-device-helper'
-import PropTypes from 'prop-types'
-
-const sDate = styles['bnk-op-date']
-const sDesc = styles['bnk-op-desc']
-const sAmount = styles['bnk-op-amount']
-const sAction = styles['bnk-op-action']
-
-class _RowDesktop extends PureComponent {
-  onSelectTransaction = () =>
-    this.props.selectTransaction(this.props.transaction)
-
-  render() {
-    const {
-      t,
-      f,
-      transaction,
-      isExtraLarge,
-      showCategoryChoice,
-      filteringOnAccount,
-      ...props
-    } = this.props
-
-    const categoryId = getCategoryId(transaction)
-    const categoryName = getCategoryName(categoryId)
-    const categoryTitle = t(`Data.subcategories.${categoryName}`)
-    const parentCategory = getParentCategory(categoryId)
-
-    const account = transaction.account.data
-
-    return (
-      <tr>
-        <td className={cx(sDesc, 'u-pv-half', 'u-pl-1')}>
-          <Media className="u-clickable">
-            <Img title={categoryTitle} onClick={showCategoryChoice}>
-              <CategoryIcon
-                category={parentCategory}
-                className={styles['bnk-op-caticon']}
-              />
-            </Img>
-            <Bd className="u-pl-1">
-              <List.Content onClick={this.onSelectTransaction}>
-                <Text>{getLabel(transaction)}</Text>
-                {!filteringOnAccount && (
-                  <Caption className={styles['bnk-op-desc-caption']}>
-                    {getAccountLabel(account)}
-                    {' - '}
-                    {getAccountInstitutionLabel(account)}
-                  </Caption>
-                )}
-              </List.Content>
-            </Bd>
-          </Media>
-        </td>
-        <TdSecondary
-          className={cx(sDate, 'u-clickable')}
-          onClick={this.onSelectTransaction}
-        >
-          {f(transaction.date, `D ${isExtraLarge ? 'MMMM' : 'MMM'} YYYY`)}
-        </TdSecondary>
-        <TdSecondary
-          className={cx(sAmount, 'u-clickable')}
-          onClick={this.onSelectTransaction}
-        >
-          <Figure
-            total={transaction.amount}
-            currency={transaction.currency}
-            coloredPositive
-            signed
-          />
-        </TdSecondary>
-        <TdSecondary className={sAction}>
-          <TransactionActions
-            transaction={transaction}
-            urls={props.urls}
-            brands={props.brands}
-            onlyDefault
-          />
-        </TdSecondary>
-      </tr>
-    )
-  }
-}
-
-export const RowDesktop = compose(
-  translate(),
-  withDispatch,
-  withUpdateCategory()
-)(_RowDesktop)
-
-class _RowMobile extends React.PureComponent {
-  render() {
-    const { transaction, t, filteringOnAccount, ...props } = this.props
-    const account = transaction.account.data
-    return (
-      <List.Row>
-        <Media className='u-full-width'>
-          <Img
-            className="u-clickable u-mr-half"
-            title={t(
-              `Data.subcategories.${getCategoryName(
-                getCategoryId(transaction)
-              )}`
-            )}
-            onClick={this.handleSelect}
-          >
-            <CategoryIcon
-              category={getParentCategory(getCategoryId(transaction))}
-            />
-          </Img>
-          <Bd className="u-clickable u-mr-half">
-            <List.Content onClick={this.handleSelect}>
-              <Text className="u-ellipsis">{getLabel(transaction)}</Text>
-              {!filteringOnAccount && (
-                <Caption
-                  className={cx('u-ellipsis', styles['bnk-op-desc-caption'])}
-                >
-                  {getAccountLabel(account)}
-                  {' - '}
-                  {getAccountInstitutionLabel(account)}
-                </Caption>
-              )}
-            </List.Content>
-          </Bd>
-          <Img onClick={this.handleSelect} className="u-clickable">
-            <Figure
-              total={transaction.amount}
-              currency={transaction.currency}
-              coloredPositive
-              signed
-            />
-          </Img>
-          <Img className={styles['bnk-transaction-mobile-action']}>
-            <TransactionActions
-              transaction={transaction}
-              urls={props.urls}
-              brands={props.brands}
-              onlyDefault
-              compact
-              menuPosition="right"
-            />
-          </Img>
-        </Media>
-      </List.Row>
-    )
-  }
-
-  handleSelect = () => {
-    this.props.selectTransaction(this.props.transaction)
-  }
-}
-
-_RowMobile.propTypes = {
-  transaction: PropTypes.object.isRequired,
-  urls: PropTypes.object.isRequired,
-  brands: PropTypes.array.isRequired
-}
-
-export const RowMobile = translate()(_RowMobile)
+import { RowDesktop, RowMobile } from './TransactionRow'
 
 const groupByDateAndSort = transactions => {
   const byDate = groupBy(transactions, x => format(x.date, 'YYYY-MM-DD'))
   return sortBy(toPairs(byDate), x => x[0]).reverse()
-}
-
-class ScrollSpy {
-  constructor(getScrollingElement) {
-    this.getScrollingElement = getScrollingElement
-    this.nodes = {}
-  }
-
-  addNode(id, node) {
-    this.nodes[id] = node
-  }
-
-  getTopMostVisibleNodeId() {
-    const scrollEl = this.getScrollingElement()
-    const topRoot =
-      scrollEl === window ? 0 : scrollEl.getBoundingClientRect().top
-    const offsets = sortBy(
-      Object.entries(this.nodes).map(([tId, node]) => [
-        tId,
-        node ? node.getBoundingClientRect().top : Infinity
-      ]),
-      x => x[1]
-    )
-    const topMost = find(offsets, o => {
-      const offset = o[1]
-      return offset - topRoot > 0
-    })
-    return topMost ? topMost[0] : null
-  }
 }
 
 const loadMoreStyle = { textAlign: 'center' }
@@ -295,7 +89,7 @@ class TransactionsD extends React.Component {
 
   constructor(props) {
     super(props)
-    this.scrollSpy = new ScrollSpy(this.getScrollingElement)
+    this.scrollSpy = new TopMost(this.getScrollingElement)
     this.handleScroll = (isIOSApp() ? debounce : throttle)(
       this.handleScroll.bind(this),
       300,
@@ -350,7 +144,6 @@ class TransactionsD extends React.Component {
 
   renderTransactions() {
     const {
-      f,
       t,
       selectTransaction,
       limitMin,
