+++ src/ducks/transactions/scroll/TopMost.js
@@ -1,0 +1,36 @@
+import sortBy from 'lodash/sortBy'
+
+/**
+ * TopMost stores DOM nodes and computes the top most
+ * visible node.
+ */
+class TopMost {
+  constructor(getScrollingElement) {
+    this.getScrollingElement = getScrollingElement
+    this.nodes = {}
+  }
+
+  addNode(id, node) {
+    this.nodes[id] = node
+  }
+
+  getTopMostVisibleNodeId() {
+    const scrollEl = this.getScrollingElement()
+    const topRoot =
+      scrollEl === window ? 0 : scrollEl.getBoundingClientRect().top
+    const offsets = sortBy(
+      Object.entries(this.nodes).map(([tId, node]) => [
+        tId,
+        node ? node.getBoundingClientRect().top : Infinity
+      ]),
+      x => x[1]
+    )
+    const topMost = find(offsets, o => {
+      const offset = o[1]
+      return offset - topRoot > 0
+    })
+    return topMost ? topMost[0] : null
+  }
+}
+
+export default TopMost
