+++ src/ducks/transactions/scroll/index.js
@@ -1,2 +1,3 @@
 export { default as ScrollRestore } from './ScrollRestore'
 export { default as InfiniteScroll } from './InfiniteScroll'
+export { default as TopMost } from './TopMost'
